Split components in review
