+++ feedback_review_frontend/src/components/mod.rs
@@ -1,0 +1,5 @@
+pub mod home;
+pub mod date;
+pub mod version;
+pub mod footer;
+pub mod not_found;
